Rename FeedAggregatorDomains to LinkAggregatorDomains
Reflects that these hosts are link aggregators: feeds whose items often
link out to other domains, rather than feeds that aggregate other feeds.
